# openapi.yaml
openapi: 3.1.0
info:
  title: Pong API
  description: |
    Pong is a health monitoring API that lets you track uptime, latency, and
    availability of your services. Create monitors for HTTP endpoints, SSL
    certificates, heartbeat checks, keyword presence, and port
    availability. Configure alert contacts to get notified when things go wrong,
    and publish public status pages so your users stay informed.

    ## Authentication

    All endpoints under `/api/v1/` require authentication via one of:

    - **Auth0 JWT** — RS256 Bearer token issued by Auth0 (used by the web app)
    - **API Key** — Long-lived key in the format `pong_<hex>` (used for scripts and CI/CD)

    Both are passed as `Authorization: Bearer <token>`.

    API keys can be created and managed via the `/api/v1/api-keys` endpoints.
    The raw key is only shown once at creation time.

    Public endpoints such as heartbeat pings and status pages do not require
    authentication.
  version: 1.0.0
servers:
  - url: https://api.getpong.dev
    description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 RS256 JSON Web Token
  schemas:
...
    StatusPage:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        name:
          type: string
          example: "Production Status"
        token:
          type: string
          description: Unguessable token used as the public URL path.
          example: "a8f3b2c1d4e5f6a7b8c9d0e1f2a3b4c5"
        description:
          type: string
          example: "Status page for production services"
        has_password:
          type: boolean
          description: Whether the page requires a password to view.
          example: false
        created_at:
          type: string
          format: date-time
          example: "2026-03-29T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-03-29T10:00:00Z"
        monitor_ids:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
    StatusPageCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Production Status"
        description:
          type: string
          example: "Status page for production services"
        password:
          type: string
          description: Optional password to protect the page. Omit or leave empty for no password.
          example: "s3cret"
        monitor_ids:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
    StatusPageUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        password:
          type: string
          description: Set a new password, send empty string to remove password, omit to keep unchanged.
        monitor_ids:
          type: array
          items:
            type: integer
    DailyUptime:
      type: object
      description: Uptime data for a single day.
      required:
        - date
        - percentage
        - total_checks
      properties:
        date:
          type: string
          format: date
          description: The date (YYYY-MM-DD).
          example: "2026-03-28"
        percentage:
          type: number
          format: double
          description: Uptime percentage for the day (0-100).
          example: 99.95
        total_checks:
          type: integer
          description: Total number of checks performed that day.
          example: 288
...
    Error:
      type: object
      description: Error response returned when a request fails.
      required:
        - error
      properties:
        error:
          type: string
          description: A human-readable error message.
          example: "monitor not found"
  parameters:
    MonitorId:
      name: id
      in: path
      required: true
      description: Unique monitor identifier.
      schema:
        type: integer
        example: 42
...
    StatusPageId:
      name: id
      in: path
      required: true
      description: Unique status page identifier.
      schema:
        type: integer
        example: 5
...
    StatusPageToken:
      name: token
      in: path
      required: true
      description: Unguessable token identifying the status page.
      schema:
        type: string
        example: "a8f3b2c1d4e5f6a7b8c9d0e1f2a3b4c5"
...
  responses:
    Unauthorized:
      description: Authentication credentials are missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "unauthorized"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "not found"
    BadRequest:
      description: The request body is invalid or missing required fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "invalid request body"
paths:
...
  /api/v1/monitors/{id}/uptime/daily:
    get:
      operationId: getDailyUptime
      tags: [Monitors]
      summary: Get daily uptime
      description: Returns day-by-day uptime data for a monitor over a specified number of days.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
        - name: days
          in: query
          description: Number of days of daily uptime data to return.
          schema:
            type: integer
            default: 90
            minimum: 1
            example: 90
      responses:
        "200":
          description: Daily uptime breakdown.
          content:
            application/json:
              schema:
                type: object
                required: [monitor_id, days, daily]
                properties:
                  monitor_id:
                    type: integer
                    description: ID of the monitor.
                    example: 42
                  days:
                    type: integer
                    description: Number of days covered.
                    example: 90
                  daily:
                    type: array
                    items:
                      $ref: "#/components/schemas/DailyUptime"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------- Alert Contacts ----------
...
  /api/v1/status-pages:
    get:
      operationId: listStatusPages
      tags: [Status Pages]
      summary: List status pages
      description: Returns all status pages belonging to the authenticated user.
      responses:
        "200":
          description: A list of status pages.
          content:
            application/json:
              schema:
                type: object
                required: [status_pages]
                properties:
                  status_pages:
                    type: array
                    items:
                      $ref: "#/components/schemas/StatusPage"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createStatusPage
      tags: [Status Pages]
      summary: Create a status page
      description: |
        Creates a new status page for the authenticated user. A unique token is
        generated automatically and used as the public URL path. An optional
        password can be set to restrict access.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusPageCreate"
      responses:
        "201":
          description: Status page created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/status-pages/{id}:
    get:
      operationId: getStatusPage
      tags: [Status Pages]
      summary: Get a status page
      description: Returns a single status page by ID.
      parameters:
        - $ref: "#/components/parameters/StatusPageId"
      responses:
        "200":
          description: The requested status page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateStatusPage
      tags: [Status Pages]
      summary: Update a status page
      description: |
        Updates an existing status page. Only the fields provided in the request
        body are changed. To set a password, include the password field. To remove
        a password, send an empty string. Omit the field to keep it unchanged.
      parameters:
        - $ref: "#/components/parameters/StatusPageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusPageUpdate"
      responses:
        "200":
          description: Status page updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteStatusPage
      tags: [Status Pages]
      summary: Delete a status page
      description: Permanently deletes a status page.
      parameters:
        - $ref: "#/components/parameters/StatusPageId"
      responses:
        "204":
          description: Status page deleted successfully.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------- Heartbeat (Public) ----------
  # ---------- API Keys ----------
...
  /status/{token}:
    get:
      operationId: getPublicStatusPage
      tags: [Status Pages]
      summary: View public status page
      description: |
        Returns an HTML status page for public viewing. If the page is password
        protected, returns a password form instead. No authentication required.
      security: []
      parameters:
        - $ref: "#/components/parameters/StatusPageToken"
      responses:
        "200":
          description: HTML status page or password form.
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Status page not found.

    post:
      operationId: submitStatusPagePassword
      tags: [Status Pages]
      summary: Submit password for protected status page
      description: |
        Submits the password for a protected status page. On success, sets a
        session cookie and redirects to the GET endpoint. On failure, re-renders
        the password form with an error message.
      security: []
      parameters:
        - $ref: "#/components/parameters/StatusPageToken"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                password:
                  type: string
      responses:
        "303":
          description: Password correct. Redirects to the status page with a session cookie.
        "200":
          description: Wrong password. Re-renders the password form with an error.
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Status page not found.

  # ---------- Admin ----------
  /api/v1/admin/stats:
    get:
      operationId: getAdminStats
      tags: [Admin]
      summary: Platform statistics
      description: Returns total users, monitors, active monitors, and waitlist count. Requires admin role.
      responses:
        "200":
          description: Platform stats.
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_users:
                    type: integer
                  total_monitors:
                    type: integer
                  active_monitors:
                    type: integer
                  waitlist_count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Not found (non-admin users receive 404).
...